Added kafka configurations
