+++ buildSrc/src/main/kotlin/rt.gradle.kts
@@ -15,6 +15,9 @@ dependencies {
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.12.+")
     implementation("io.github.microutils:kotlin-logging:1.12.5")
 
+    implementation("org.springframework.kafka:spring-kafka:2.7.0")
+    implementation("org.springframework.kafka:spring-kafka-test:2.7.0")
+
     testImplementation("io.mockk:mockk:${mockkVersion}")
     testImplementation(kotlin("test-junit5"))
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.6.0")
